pallet_collective_proxy::pallet::dispatchables

Function execute_call

pub fn execute_call<T: Config>(call: Box<<T as Config>::RuntimeCall>)

Executes the call on a behalf of an aliased account.

The origin of the call is supposed to be a collective (but can be anything) which can dispatch call on behalf of the aliased account. It’s essentially a proxy call that can be made by arbitrary origin type.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::execute_call.