Trait pallet_collective_proxy::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + GetDispatchInfo + From<Call<Self>> + IsType<<Self as Config>::RuntimeCall>;
    type CollectiveProxy: EnsureOrigin<<Self as Config>::RuntimeOrigin>;
    type ProxyAccountId: Get<Self::AccountId>;
    type CallFilter: InstanceFilter<<Self as Config>::RuntimeCall> + Default;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + GetDispatchInfo + From<Call<Self>> + IsType<<Self as Config>::RuntimeCall>

The overarching call type.

type CollectiveProxy: EnsureOrigin<<Self as Config>::RuntimeOrigin>

Origin that can act on behalf of the collective.

type ProxyAccountId: Get<Self::AccountId>

Account representing the collective treasury.

type CallFilter: InstanceFilter<<Self as Config>::RuntimeCall> + Default

Filter to determine whether a call can be executed or not.

type WeightInfo: WeightInfo

Weight info